requirements.txt
customtkinter==5.2.0
Pillow>=9.0.0
psutil>=5.9.0
watchdog>=3.0.0
scapy>=2.5.0
pefile>=2023.2.7
scikit-learn>=1.2.0
matplotlib>=3.7.0
numpy>=1.24.0
cryptography>=41.0.0
joblib>=1.3.0
packaging>=23.0